Move ResourceTypeSelect
